requirements.txt
requests
dnspython
tldextract
urllib3